requirements.txt
bilibili-api-python>=16.0.0
yt-dlp>=2024.0.0
funasr>=1.0.0
anthropic>=0.30.0
openai>=1.0.0
click>=8.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
rich>=13.0.0
pydantic>=2.0.0
aiohttp>=3.9.0
pymupdf>=1.23.0
python-docx>=1.0.0
